:art: re-organise landing page files
