+++ src/pages/landing-page/subpages/smartpanel/constants.ts
