+++ src/pages/landing-page/subpages/smartpanel/utils.ts
